% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dt.R
\name{mod_dt_server}
\alias{mod_dt_server}
\title{export Server Function}
\usage{
mod_dt_server(
  id = NULL,
  data,
  output_id = "dt",
  scroll_y = 300,
  left = 1L,
  dt_bundle_buttons = dtf::dt_bundle_buttons_en,
  dt_bundle_internationalization = dtf::dt_bundle_internationalization_en,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{dt_bundle_buttons}{[\link{function}] Seet [\link[dtf:dt_bundle_buttons]{dtf::dt_bundle_buttons}]}

\item{dt_bundle_internationalization}{[\link{function}] Seet [\link[dtf:dt_bundle_internationalization]{dtf::dt_bundle_internationalization}]}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
export Server Function
}
