% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dt.R
\name{mod_dt_ui}
\alias{mod_dt_ui}
\title{export UI Function}
\usage{
mod_dt_ui(
  id = "dt",
  select_ui = tagList(),
  data_title = "Data table",
  data_width = 12,
  outer_box = FALSE,
  outer_title = "Select columns",
  outer_width = 12,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
A shiny Module.
}
