% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_freq_table.R
\name{mod_eda_freq_table_server}
\alias{mod_eda_freq_table_server}
\title{export Server Function}
\usage{
mod_eda_freq_table_server(id = NULL, r_data)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{r_data}{[\link{reactive}] Reactive function that serves data}
}
\description{
export Server Function
}
