% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_freq_table.R
\name{mod_eda_freq_table_ui}
\alias{mod_eda_freq_table_ui}
\title{export UI Function}
\usage{
mod_eda_freq_table_ui(
  id = NULL,
  grouping_title = "Group data",
  grouping_width = 12,
  grouping_button_label = "Add group-by statement",
  grouping_button_class = "btn-success",
  grouping_button_style = "color: #fff;",
  grouping_button_icon = icon("plus"),
  grouping_button_width = 190,
  freq_title = "Frequency table",
  freq_width = 12,
  outer_box = FALSE,
  outer_title = "Frequency table",
  outer_width = 12
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{outer_box}{[\link{logical}] Wrap in outer box yes/no}

\item{outer_width}{}
}
\description{
A shiny Module.
}
