% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_select.R
\name{mod_eda_select_server}
\alias{mod_eda_select_server}
\title{export Server Function}
\usage{
mod_eda_select_server(
  id = "eda_select",
  r_data,
  input_id_prefix = "select_input"
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{input_id_prefix}{}
}
\description{
export Server Function
}
