% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_freq_table3.R
\name{mod_freq_table3_server}
\alias{mod_freq_table3_server}
\title{export Server Function}
\usage{
mod_freq_table3_server(
  id = "freq_table",
  data,
  render_table = FALSE,
  dt_bundle_buttons = dtf::dt_bundle_buttons_en,
  dt_bundle_internationalization = dtf::dt_bundle_internationalization_en,
  dt_bundles_additional = list(),
  transform_fn = identity,
  dt.filter = dtf::valid_dt_filter_values(1),
  dt.scroll_y = 300,
  dt.left = 1L,
  input_id_prefix = "grouping_input"
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{data}{[\link{reactive}] Reactive function that serves data}

\item{dt_bundle_buttons}{[\link{function}] Seet [\link[dtf:dt_bundle_buttons]{dtf::dt_bundle_buttons}]}

\item{dt_bundle_internationalization}{[\link{function}] Seet [\link[dtf:dt_bundle_internationalization]{dtf::dt_bundle_internationalization}]}

\item{dt_bundles_additional}{[\link{list}] Additional DT bundles. See
\link[dtf:dt_bundle_dom]{dtf::dt_bundle_dom} as an example}

\item{dt.filter}{[\link{character}] DT column filter settings. See package \link{dtf}}

\item{dt.scroll_y}{[\link{integer}] DT: pixels for vertical scrolling}

\item{dt.left}{[\link{integer}] DT: columns to fix from the left}

\item{input_id_prefix}{}
}
\description{
export Server Function
}
