% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_freq_table.R
\name{mod_freq_table_ui}
\alias{mod_freq_table_ui}
\title{export UI Function}
\usage{
mod_freq_table_ui(
  id = "freq_table",
  ns = function() {
 },
  grouping_title = "Group by",
  grouping_width = 12,
  grouping_button_label = "Add group-by statement",
  grouping_button_class = "btn-primary",
  grouping_button_icon = icon("plus"),
  grouping_button_width = 190,
  grouping.button_submit_label = "Submit",
  grouping.button_submit_class = "btn-success",
  grouping.button_submit_style = "color: #fff;",
  grouping.button_submit_icon = icon("check"),
  grouping.button_submit_width = 100,
  freq.box_title = "Frequency table",
  freq.box_width = 12,
  freq.box_title_width = 200,
  freq.dropdown_title = "Column names",
  freq.dropdown_width = 150,
  freq.dropdown_col_n_abs_label = "Name for abs. count",
  freq.dropdown_col_n_rel_label = "Name for rel. count",
  freq.col_n_abs = "n_abs",
  freq.col_n_rel = "n_rel",
  outer_box = FALSE,
  outer_title = "Frequency table",
  outer_width = 12,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{ns}{[\link{function}] Namespace function}

\item{grouping.button_submit_width}{}

\item{outer_box}{[\link{logical}] Wrap in outer box yes/no}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
A shiny Module.
}
