% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_group_by2.R
\name{mod_group_by2_ui}
\alias{mod_group_by2_ui}
\title{export UI Function}
\usage{
mod_group_by2_ui(
  id = "freq_table",
  ns = function() {
 },
  grouping_title = "Group by",
  grouping_width = 12,
  grouping.button_submit_label = "Submit",
  grouping.button_submit_class = "btn-success",
  grouping.button_submit_style = "color: #fff;",
  grouping.button_submit_icon = icon("check"),
  grouping.button_submit_width = 100,
  outer_box = FALSE,
  outer_title = "Frequency table",
  outer_width = 12,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{ns}{[\link{function}] Namespace function}

\item{grouping.button_submit_width}{}

\item{outer_box}{[\link{logical}] Wrap in outer box yes/no}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
A shiny Module.
}
