% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select2.R
\name{mod_select2_server}
\alias{mod_select2_server}
\title{export Server Function}
\usage{
mod_select2_server(
  id = "eda_select",
  data,
  input_id_prefix = "selectize_input",
  dt_bundle_buttons = dtf::dt_bundle_buttons_en,
  dt_bundle_internationalization = dtf::dt_bundle_internationalization_en,
  verbose = FALSE,
  fn_select_ui = create_selectize_ui,
  render_data = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{dt_bundle_buttons}{[\link{function}] Seet [\link[dtf:dt_bundle_buttons]{dtf::dt_bundle_buttons}]}

\item{dt_bundle_internationalization}{[\link{function}] Seet [\link[dtf:dt_bundle_internationalization]{dtf::dt_bundle_internationalization}]}

\item{verbose}{[\link{logical}] Tracing infos yes/no}

\item{render_data}{[\link{logical}] Render data table yes/no}
}
\description{
export Server Function
}
