% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_old2.R
\name{mod_select2_ui}
\alias{mod_select2_ui}
\title{export UI Function}
\usage{
mod_select2_ui(
  id = "eda_select",
  ns = function() {
 },
  select_title = "Column filter (none means 'all columns')",
  select_width = 12,
  select_id = "select_ui",
  select_button_label = "Submit",
  select_button_class = "btn-success",
  select_button_style = "color: #fff;",
  select_button_icon = icon("check"),
  select_button_width = 100,
  render_data = FALSE,
  data_title = "Data table",
  data_width = 12,
  outer_box = FALSE,
  outer_title = "Select columns",
  outer_width = 12,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{ns}{[\link{function}] Namespace function}

\item{render_data}{[\link{logical}] Render data table yes/no}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
A shiny Module.
}
