% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_dt.R
\name{mod_table_dt_server}
\alias{mod_table_dt_server}
\title{export Server Function}
\usage{
mod_table_dt_server(
  id = "table_dt",
  output_id = "dt",
  data,
  filter = dtf::valid_dt_filter_values(1),
  scrollY = 400,
  left = integer(),
  right = integer(),
  server = FALSE,
  trans_fn = identity,
  rename_fn = identity,
  .bundles = list(),
  .rownames = TRUE,
  .editable = FALSE,
  .escape = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{verbose}{[\link{logical}] Tracing infos yes/no}

\item{...}{}
}
\description{
export Server Function
}
