% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_dt.R
\name{mod_table_dt_ui}
\alias{mod_table_dt_ui}
\title{export UI Function}
\usage{
mod_table_dt_ui(
  id = "table_dt",
  ns = function() {
 },
  output_id = "dt",
  box = FALSE,
  box_title = "Data table",
  box_width = 12,
  verbose = FALSE
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{ns}{[\link{function}] Namespace function}

\item{output_id}{[\link{character}] Output ID}

\item{verbose}{[\link{logical}] Tracing infos yes/no}
}
\description{
A shiny Module.
}
