% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reactive_dplyr.R
\name{reactive_select}
\alias{reactive_select}
\title{Reactive select}
\usage{
reactive_select(
  id = NULL,
  data,
  event_id = "submit",
  input_ids,
  input_values,
  dt_bundle_buttons = dtf::dt_bundle_buttons_en,
  dt_bundle_internationalization = dtf::dt_bundle_internationalization_en,
  ignore_null = TRUE
)
}
\arguments{
\item{ignore_null}{}
}
\description{
Reactive select
}
