% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_old2.R
\name{render_select_data_table2}
\alias{render_select_data_table2}
\title{Title}
\usage{
render_select_data_table2(
  id = NULL,
  data,
  event_id = "submit",
  dt_bundle_buttons = dtf::dt_bundle_buttons_en,
  dt_bundle_internationalization = dtf::dt_bundle_internationalization_en
)
}
\arguments{
\item{buttons_language}{}
}
\description{
Title
}
