% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimo.eda-package.R
\docType{package}
\name{shimo.eda-package}
\alias{shimo.eda}
\alias{shimo.eda-package}
\title{shimo.eda: Shiny modules: EDA}
\description{
Shiny modules for EDA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rappster/shimo.eda}
  \item Report bugs at \url{https://github.com/rappster/shimo.eda/issues}
}

}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
