% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_shiny.R
\name{shiny_font_size_perc}
\alias{shiny_font_size_perc}
\title{Adjust font size (in percent)}
\usage{
shiny_font_size_perc(value, size = 75)
}
\arguments{
\item{value}{[\link{character}] Text value}

\item{size}{[\link{integer}] Font siz in percent}
}
\value{
\link[shiny:reexports]{shiny::div}
}
\description{
Adjust font size (in percent)
}
\examples{
shiny_font_size_perc("foo")
shiny_font_size_perc("foo", 50)
}
