% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{tabitem_vertical_space}
\alias{tabitem_vertical_space}
\title{Add vertical space for \code{tabItem}}
\usage{
tabitem_vertical_space(times = 3)
}
\arguments{
\item{times}{[\link{integer}] Number of \code{tag$br()} to set}
}
\description{
Useful when using \href{https://github.com/rstudio/shinydashboard}{shinyDashboard}
}
\examples{
tabitem_vertical_space()
tabitem_vertical_space(2)
}
