% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{vertical_space}
\alias{vertical_space}
\title{Add vertical space}
\usage{
vertical_space(times = 1)
}
\arguments{
\item{times}{[\link{integer}] Number of \code{tag$br()} to set}
}
\description{
Useful when using \href{https://github.com/rstudio/shinydashboard}{shinyDashboard}
}
\examples{
vertical_space()
vertical_space(2)
}
