% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_freq_table.R
\name{mod_eda_freq_table_server}
\alias{mod_eda_freq_table_server}
\title{export Server Function}
\usage{
mod_eda_freq_table_server(
  id = NULL,
  r_data,
  dt_bundle_buttons = dti::dt_bundle_buttons_en,
  dt_bundle_internationalization = dti::dt_bundle_internationalization_en
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{r_data}{[\link{reactive}] Reactive function that serves data}

\item{dt_bundle_buttons}{[\link{function}] Seet [\link[dti:dt_bundle_buttons]{dti::dt_bundle_buttons}]}

\item{dt_bundle_internationalization}{[\link{function}] Seet [\link[dti:dt_bundle_internationalization]{dti::dt_bundle_internationalization}]}
}
\description{
export Server Function
}
