% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_select.R
\name{mod_eda_select_server}
\alias{mod_eda_select_server}
\title{export Server Function}
\usage{
mod_eda_select_server(
  id = "eda_select",
  r_data,
  input_id_prefix = "select_input",
  dt_bundle_buttons = dti::dt_bundle_buttons_en,
  dt_bundle_internationalization = dti::dt_bundle_internationalization_en
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{dt_bundle_buttons}{[\link{function}] Seet [\link[dti:dt_bundle_buttons]{dti::dt_bundle_buttons}]}

\item{dt_bundle_internationalization}{[\link{function}] Seet [\link[dti:dt_bundle_internationalization]{dti::dt_bundle_internationalization}]}
}
\description{
export Server Function
}
