% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_select.R
\name{mod_eda_select_ui}
\alias{mod_eda_select_ui}
\title{export UI Function}
\usage{
mod_eda_select_ui(
  id = "eda_select",
  select_title = "Select columns",
  select_width = 12,
  select_id = "select_ui",
  select_button_label = "Add select statement",
  select_button_class = "btn-success",
  select_button_style = "color: #fff;",
  select_button_icon = icon("plus"),
  select_button_width = 170,
  data_title = "Data table",
  data_width = 12,
  outer_box = FALSE,
  outer_title = "Frequency table",
  outer_width = 12
)
}
\arguments{
\item{id}{[\link{character}] Module ID}

\item{outer_width}{}
}
\description{
A shiny Module.
}
