% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_freq_table.R
\name{render_grouping_data_table}
\alias{render_grouping_data_table}
\title{Title}
\usage{
render_grouping_data_table(
  id = NULL,
  r_data,
  input_ids,
  input_values,
  output_id = "grouping_tbl",
  freq_tab.col_n_abs = "col_n_abs",
  freq_tab.col_n_rel = "col_n_rel",
  freq_tab.sort = TRUE,
  dt_bundle_buttons = dti::dt_bundle_buttons_en,
  dt_bundle_internationalization = dti::dt_bundle_internationalization_en
)
}
\arguments{
\item{dt_bundle_internationalization}{}
}
\description{
Title
}
