% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eda_select.R
\name{render_select_data_table}
\alias{render_select_data_table}
\title{Title}
\usage{
render_select_data_table(
  id = NULL,
  r_data,
  input_ids,
  input_values,
  dt_bundle_buttons = dti::dt_bundle_buttons_en,
  dt_bundle_internationalization = dti::dt_bundle_internationalization_en
)
}
\arguments{
\item{buttons_language}{}
}
\description{
Title
}
